move component to its own file
